+++ graphics/src/renderer.rs
@@ -135,14 +135,14 @@ where
         self.backend().default_font()
     }
 
-    fn default_size(&self) -> u16 {
+    fn default_size(&self) -> f32 {
         self.backend().default_size()
     }
 
     fn measure(
         &self,
         content: &str,
-        size: u16,
+        size: f32,
         font: Font,
         bounds: Size,
     ) -> (f32, f32) {
